+++ billy/web/admin/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-class BillyAdminConfig(AppConfig):
-    name = 'billy.web.admin'
-    label = 'billyadmin'
-    verbose_name = "Billy Admin"
-