+++ billy/web/api/tests/__init__.py
@@ -7,6 +7,5 @@ from .test_metadata import *
 from .test_bills import *
 from .test_legislators import *
 from .test_committees import *
-from .test_events import *
 from .test_districts import *
 from .test_boundaries import *